pyproject.toml
[tool.poetry]
name = "newspipe"
version = "8.0.0"
description = "Newspipe is a web-based news aggregator and reader."
authors = ["Cédric Bonhomme <cedric@cedricbonhomme.org>"]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.6.2"
requests = "^2.23.0"
chardet = "^3.0.4"
requests-futures = "^1.0.0"
feedparser = "^5.2.1"
beautifulsoup4 = "^4.8.2"
lxml = "^4.5.0"
opml = "^0.5"
SQLAlchemy = "^1.3.13"
alembic = "^1.4.0"
Flask = "^1.1.1"
Flask-SQLAlchemy = "^2.4.1"
Flask-Login = "^0.5.0"
Flask-Principal = "^0.4.0"
Flask-WTF = "^0.14.3"
Flask-RESTful = "^0.3.8"
Flask-Restless = "^0.17.0"
Flask-paginate = "^0.5.5"
Flask-Babel = "^1.0.0"
Flask-SSLify = "^0.1.5"
Flask-Migrate = "^2.5.2"
Flask-Script = "^2.0.6"
WTForms = "^2.2.1"
sendgrid = "^6.1.2"
python-dateutil = "^2.8.1"
psycopg2-binary = "^2.8.4"



[tool.poetry.dev-dependencies]
sqlalchemy-schemadisplay = "^1.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
